requirements_prod.txt
Django==3.1.5
django-rest-authtoken==2.1.3
djangorestframework==3.12.2
mysql-connector-python==8.0.22
mysqlclient==2.0.1
requests==2.25.1
urllib3==1.26.3
#wheel==0.36.2
#build==0.3.1.post1
gunicorn==20.1.0